+++ plugins/resources/compound/libcompound.cpp
@@ -662,7 +662,7 @@ extern "C" {
     /// @brief interface for POSIX lseek
     eirods::error compound_file_lseek(
         eirods::resource_operation_context* _ctx,
-        size_t                              _offset, 
+        long long                           _offset, 
         int                                 _whence ) {
         // =-=-=-=-=-=-=-
         // check the context for validity
@@ -681,7 +681,7 @@ extern "C" {
 
         // =-=-=-=-=-=-=-
         // forward the call
-        return resc->call< size_t, int >( _ctx->comm(), eirods::RESOURCE_OP_LSEEK, _ctx->fco(), _offset, _whence );
+        return resc->call< long long, int >( _ctx->comm(), eirods::RESOURCE_OP_LSEEK, _ctx->fco(), _offset, _whence );
  
     } // compound_file_lseek
